@@ -1,5 +1,6 @@
 project(DFA-Creator)
-set(CMAKE_CXX_STANDARD 11)
+cmake_minimum_required(VERSION 3.0)
+set(CMAKE_CXX_STANDARD 14)
 
 file(GLOB srcs *.cpp *.h)
 
